requirements.txt
streamlit==1.13.1
torch==1.12.1
torchvision==0.13.0
Pillow
matplotlib
